Service layer and user creation controller and execute completed transaction

+++ src/main/java/com/transaction/usercase/dto/TransactionRequest.java
@@ -1,6 +1,8 @@
 package com.transaction.usercase.dto;
 
-public record TransactionRequest(Double value,
+import java.math.BigDecimal;
+
+public record TransactionRequest(BigDecimal value,
                                  String payer,
                                  String payee) {
 }
